+++ src/util.cpp
@@ -1,8 +1,8 @@
 #include "dev_or_prod.h"
 
-uint32_t getCurrentTimeMs() {
+unsigned long getCurrentTimeMs() {
   TickType_t currentTick = xTaskGetTickCount();
-  uint32_t currentTimeMs = currentTick * portTICK_PERIOD_MS;
+  unsigned long currentTimeMs = currentTick * portTICK_PERIOD_MS;
   return currentTimeMs;
 }
